Restructure impl GasHandler, and use it in runtime
