+++ runtime/src/lib.rs
@@ -23,16 +23,14 @@
 
 use cennznet_primitives::{
 	traits::BuyFeeAsset,
-	types::{AccountId, AccountIndex, AssetId, Balance, BlockNumber, FeeExchange, Hash, Index, Moment, Signature},
+	types::{AccountId, AccountIndex, AssetId, Balance, BlockNumber, Hash, Index, Moment, Signature},
 };
 use cennznut::{CENNZnut, Domain, Validate, ValidationErr};
 use codec::Decode;
 use frame_support::{
-	additional_traits, construct_runtime, debug, parameter_types, storage::unhashed, traits::Randomness,
-	weights::Weight,
+	additional_traits, construct_runtime, debug, parameter_types, traits::Randomness, weights::Weight,
 };
 use frame_system::offchain::TransactionSubmitter;
-use pallet_contracts::GasMeter;
 use pallet_contracts_rpc_runtime_api::ContractExecResult;
 use pallet_generic_asset::{SpendingAssetCurrency, StakingAssetCurrency};
 use pallet_grandpa::fg_primitives;
@@ -74,11 +72,11 @@ pub use crml_sylo::vault as sylo_vault;
 
 /// Implementations of some helper traits passed into runtime modules as associated types.
 pub mod impls;
-use impls::{CurrencyToVoteHandler, FeeMultiplierUpdateHandler, LinearWeightToFee};
+use impls::{CurrencyToVoteHandler, FeeMultiplierUpdateHandler, GasHandler, LinearWeightToFee};
 
 /// Constant values used within the runtime.
 pub mod constants;
-use constants::{contract::GAS_FEE_EXCHANGE_KEY, currency::*, time::*};
+use constants::{currency::*, time::*};
 
 mod doughnut;
 
@@ -452,7 +450,7 @@ impl pallet_contracts::Trait for Runtime {
 	type ComputeDispatchFee = pallet_contracts::DefaultDispatchFeeComputor<Runtime>;
 	type TrieIdGenerator = pallet_contracts::TrieIdFromParentCounter<Runtime>;
 	type GasPayment = ();
-	type GasHandler = ();
+	type GasHandler = GasHandler;
 	type RentPayment = ();
 	type SignedClaimHandicap = pallet_contracts::DefaultSignedClaimHandicap;
 	type TombstoneDeposit = TombstoneDeposit;
@@ -585,30 +583,6 @@ impl additional_traits::DelegatedDispatchVerifier for Runtime {
 			Err(ValidationErr::NoPermission(Domain::MethodArguments)) => {
 				Err("CENNZnut does not grant permission for method arguments")
 			}
-		}
-	}
-}
-
-/// Handles gas payment post contract execution (before deferring runtime calls) via CENNZX-Spot exchange.
-impl<T> pallet_contracts::GasHandler<T> for Runtime
-where
-	T: pallet_contracts::Trait + pallet_generic_asset::Trait + crml_cennzx_spot::Trait + Decode,
-{
-	fn empty_unused_gas(transactor: &T::AccountId, gas_meter: GasMeter<T>) {
-		// The take() function ensures the entry is `killed` after access.
-		// Note: if there's no entry (ie, None), it means we handled an edge case of
-		// CPAY in FeeExchange, ie, gas is already paid in CPAY.
-		if let Some(exchange_op) = unhashed::take::<FeeExchange>(&GAS_FEE_EXCHANGE_KEY) {
-			let gas_spent = gas_meter.spent();
-
-			// Fee exchange can never fail as conditions such as having enough liquidity
-			// are checked early (before FeeExchange is put into storage)
-			let _ = <crml_cennzx_spot::Module<T> as BuyFeeAsset<_, _>>::buy_fee_asset(
-				transactor,
-				gas_spent.saturated_into(),
-				&exchange_op,
-			)
-			.unwrap();
 		}
 	}
 }
